requirements.txt
Cython
packaging
nemo_toolkit[asr]
omegaconf
soundfile
numpy
